refactor: ResultMark 를 view 로 이동
